requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dnspython==2.8.0
dotenv==0.9.9
et_xmlfile==2.0.0
fastapi==0.135.2
greenlet==3.3.1
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
openpyxl==3.1.5
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pymongo==4.16.0
pytest==9.0.2
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.2
regex==2026.1.15
requests==2.33.0
soupsieve==2.8.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
itsdangerous
